Re-organized packages and made snake-case
